added Policy to Hobby Tag and User for Authorisation

+++ app/Http/Controllers/HobbyController.php
@@ -7,6 +7,7 @@ use App\Tag;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Session;
 use Intervention\Image\Facades\Image;
+use Illuminate\Support\Facades\Gate;
 
 class HobbyController extends Controller
 {
@@ -97,6 +98,8 @@ class HobbyController extends Controller
      */
     public function edit(Hobby $hobby)
     {
+        abort_unless(Gate::allows('update', $hobby), 403);
+
         return view('hobby.edit', [
             'hobby' => $hobby,
             'message_success' => Session::get('message_success'),
@@ -113,6 +116,8 @@ class HobbyController extends Controller
      */
     public function update(Request $request, Hobby $hobby)
     {
+        abort_unless(Gate::allows('update', $hobby), 403); // HobbyPolicy applied here.
+
         $request->validate([
             'name' => 'required|min:3',
             'description' => 'required|min:5',
@@ -141,6 +146,8 @@ class HobbyController extends Controller
      */
     public function destroy(Hobby $hobby)
     {
+        abort_unless(Gate::allows('delete', $hobby), 403); // HobbyPolicy applied here.
+
         $old_name = $hobby->name;
         $hobby->delete();
         return $this->index()->with([
